Add files via upload
The Employee Performance and Recognition Portal are to replace the existing manual system with a software solution. 
It allows all the employees in different sections of the company to post rating details individually which will be forwarded to their appraiser. 
It allows the appraiser to re-rate the employee individual task ratings which takes into consideration and calculates average rating of the all employees based on no of tasks they have done, their ratings and weightage. 
Based on these average ratings this system will generate consolidated ranks for the employee automatically with any errors.